Do some refactoring
