+++ ui/components/multichain-accounts/hardware-wallets/ledger/index.ts
@@ -5,11 +5,13 @@ export {
   LEDGER_CONNECTION_STATUS_ILLUSTRATION_URL,
 } from './ledger-connection-status';
 export type {
+  LedgerConnectionStatusDeviceNotFound,
   LedgerConnectionStatusProps,
   LedgerConnectionStatusType,
 } from './ledger-connection-status';
 export { LedgerConnectionStatusDeviceSelector } from './ledger-connection-status-device-selector';
 export type { LedgerConnectionStatusDeviceSelectorProps } from './ledger-connection-status-device-selector';
 export { LedgerConnectionStatusIllustration } from './ledger-connection-status-illustration';
+export type { LedgerConnectionStatusIllustrationProps } from './ledger-connection-status-illustration';
 export { LedgerConnectionStatusInstructions } from './ledger-connection-status-instructions';
 export type { LedgerConnectionStatusInstructionsProps } from './ledger-connection-status-instructions';
