+++ ui/components/multichain-accounts/hardware-wallets/ledger/ledger-connection-status-illustration/index.ts
@@ -1,1 +1,2 @@
 export { LedgerConnectionStatusIllustration } from './ledger-connection-status-illustration';
+export type { LedgerConnectionStatusIllustrationProps } from './ledger-connection-status-illustration.types';
