+++ ui/components/multichain-accounts/hardware-wallets/ledger/ledger-connection-status-illustration/ledger-connection-status-illustration.stories.tsx
@@ -5,7 +5,7 @@ import {
   BoxBackgroundColor,
   BoxFlexDirection,
 } from '@metamask/design-system-react';
-import { LEDGER_CONNECTION_STATUS } from '../ledger-connection-status';
+import { LEDGER_CONNECTION_STATUS } from '../ledger-connection-status.constants';
 import { LedgerConnectionStatusIllustration } from '.';
 
 const STORY_FRAME_WIDTH = 460;
