+++ ui/components/multichain-accounts/hardware-wallets/ledger/ledger-connection-status-illustration/ledger-connection-status-illustration.test.tsx
@@ -8,7 +8,7 @@ import {
 import {
   LEDGER_CONNECTION_STATUS_ILLUSTRATION_URL,
   type LedgerConnectionStatusType,
-} from '../ledger-connection-status';
+} from '../ledger-connection-status.constants';
 import { LedgerConnectionStatusIllustration } from '.';
 
 describe('LedgerConnectionStatusIllustration', () => {
