+++ ui/components/multichain-accounts/hardware-wallets/ledger/ledger-connection-status-illustration/ledger-connection-status-illustration.types.ts
@@ -1,0 +1,6 @@
+import type { LedgerConnectionStatusProps } from '../ledger-connection-status/ledger-connection-status.types';
+
+export type LedgerConnectionStatusIllustrationProps = Pick<
+  LedgerConnectionStatusProps,
+  'status'
+>;
