+++ ui/components/multichain-accounts/hardware-wallets/ledger/ledger-connection-status-instructions/ledger-connection-status-instructions.types.ts
@@ -1,6 +1,6 @@
-import { LEDGER_CONNECTION_STATUS } from '../ledger-connection-status/ledger-connection-status.constants';
+import type { LedgerConnectionStatusDeviceNotFound } from '../ledger-connection-status.constants';
 
 export type LedgerConnectionStatusInstructionsProps = {
   /** Must be `device-not-found`; renders the troubleshooting checklist for that state. */
-  status: typeof LEDGER_CONNECTION_STATUS.DeviceNotFound;
+  status: LedgerConnectionStatusDeviceNotFound;
 };
