+++ ui/components/multichain-accounts/hardware-wallets/ledger/ledger-connection-status/index.ts
@@ -3,6 +3,9 @@ export {
   LEDGER_CONNECTION_STATUS,
   LEDGER_CONNECTION_STATUS_CONTENT,
   LEDGER_CONNECTION_STATUS_ILLUSTRATION_URL,
-} from './ledger-connection-status.constants';
-export type { LedgerConnectionStatusType } from './ledger-connection-status.constants';
+} from '../ledger-connection-status.constants';
+export type {
+  LedgerConnectionStatusDeviceNotFound,
+  LedgerConnectionStatusType,
+} from '../ledger-connection-status.constants';
 export type { LedgerConnectionStatusProps } from './ledger-connection-status.types';
